refactor: drop application imports of infra
